# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University Management API",
        "description": "REST API for the Laravel 12 University Management System. All protected endpoints require a Sanctum Bearer token obtained from POST /api/auth/login.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/classrooms": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "List all classrooms",
                "operationId": "14d67cf1dea4ca3281347d669b2c3abf",
                "responses": {
                    "200": {
                        "description": "Array of classrooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Classroom"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Create a new classroom",
                "operationId": "3c3dd8db4436d352308a71d12d54918d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_number",
                                    "building"
                                ],
                                "properties": {
                                    "room_number": {
                                        "type": "string",
                                        "example": "A101"
                                    },
                                    "building": {
                                        "type": "string",
                                        "example": "Engineering Block"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Classroom created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Classroom"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/classrooms/{id}": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Get a classroom by ID",
                "operationId": "e9fc308e19140674ce17a5e7001fd119",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Classroom with courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Classroom"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Update a classroom",
                "operationId": "28293f9f116f21c3ae8b61d1fe14b817",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "room_number": {
                                        "type": "string",
                                        "example": "B202"
                                    },
                                    "building": {
                                        "type": "string",
                                        "example": "Science Block"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Delete a classroom",
                "operationId": "21027b559d3c63d479f07ce97b91ebdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "deleted": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Classroom": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "room_number": {
                        "type": "string",
                        "example": "A101"
                    },
                    "building": {
                        "type": "string",
                        "example": "Engineering Block"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication. Obtain a token via POST /api/auth/login then enter it here (without the \"Bearer \" prefix — Swagger UI adds that automatically).",
                "scheme": "bearer",
                "bearerFormat": "token"
            }
        }
    },
...
}